+++ CHANGELOG.md
@@ -6,11 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Added
 ### Changed 
-- [Tobii Provider] IsLive now reflects state of driver
-- [Tobii Provider] Only show Tobii devices in ProviderReport that are currently connected
 ### Deprecated
 ### Removed
 ### Fixed
+
+## 0.10.9 - 2019-12-13
+### Changed 
+- [Tobii Provider] IsLive now reflects state of driver
+- [Tobii Provider] Only show Tobii devices in ProviderReport that are currently connected
 
 ## 0.10.8 - 2019-12-13
 ### Changed